requirements.txt
beautifulsoup4
reportlab
pinecone[grpc] 
langchain-pinecone 
langchain_community 
langchain_openai 
langchain_experimental
pypdf
rich
PyMuPDF
sentence-transformers

anthropic

google-generativeai
streamlit